从gitlab移植到github

+++ src/main/java/com/example/cinema/po/MovieTotalBoxOffice.java
@@ -5,14 +5,23 @@ package com.example.cinema.po;
  * @date 2019/4/28 6:09 PM
  */
 public class MovieTotalBoxOffice {
-    private Integer movieId;
+
+    /**
+     * 电影id
+     */
+    private int movieId;
+
     /**
      * 票房(单位：元)，(PS:如果后续数据量大，可自行处理单位，如改成单位：万元)
      */
-    private Integer boxOffice;
+    private Double boxOffice;
+
+    /**
+     * 电影名
+     */
     private String name;
 
-    public Integer getMovieId() {
+    public int getMovieId() {
         return movieId;
     }
 
@@ -20,11 +29,11 @@ public class MovieTotalBoxOffice {
         this.movieId = movieId;
     }
 
-    public Integer getBoxOffice() {
+    public Double getBoxOffice() {
         return boxOffice;
     }
 
-    public void setBoxOffice(Integer boxOffice) {
+    public void setBoxOffice(Double boxOffice) {
         this.boxOffice = boxOffice;
     }
